New features; command groups, backlog limits etc...
Added an arbitrary limit to the backlog for running programs. Also added
a command to set the limit, and purge the backlog.

Fixed a deadlock potential problem for commands that output a ton of
text; this could cause the 'stdout' to block, which would also block the
receive end (same thread). Now use a separate thread to run the
commands, which allows the 'receive' end to continue safely.

Added 'command groups' -- you can add an optional 'kind' to your
command, and you can also set the command parameter per kind. This
allows multiple command groups to cohabitate (ie libraries with their
own commands).

Signed-off-by: Michel Pollet <buserror@gmail.com>